Conflicts:
	utf8.c

+++ cmd-show-messages.c
@@ -80,8 +80,8 @@ cmd_show_messages_jobs(struct cmdq_item *item, int blank)
 			cmdq_print(item, "%s", "");
 			blank = 0;
 		}
-		cmdq_print(item, "Job %u: %s [fd=%d, pid=%d, status=%d]",
-		    n, job->cmd, job->fd, job->pid, job->status);
+		cmdq_print(item, "Job %u: %s [fd=%d, pid=%ld, status=%d]",
+		    n, job->cmd, job->fd, (long)job->pid, job->status);
 		n++;
 	}
 	return (n != 0);
